app/requirements.txt
pandas==1.4.3
google-cloud-bigquery==3.2.0
google-cloud-logging==3.1.2
google-cloud-pubsub==2.13.0
google-cloud-storage==2.4.0
pandas-gbq==0.17.6